% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotTxPathway}
\alias{PlotTxPathway}
\title{Plot the outcome incidence over time}
\usage{
PlotTxPathway(
  cohortDatabaseSchema,
  cohortTable,
  cohortId,
  conceptSets,
  drugExposureData,
  sequenceData,
  outputFolder,
  StartDays = 0,
  EndDays = 365,
  pathLevel = 3
)
}
\arguments{
\item{cohortDatabaseSchema}{The plpData object returned by running getPlpData()}

\item{cohortTable}{The cohort id corresponding to the outcome}

\item{cohortId}{Remove patients who have had the outcome before their target cohort index date from the plot}

\item{conceptSets}{(integer) The time-at-risk starts at target cohort index date plus this value}

\item{drugExposureData}{(integer) The time-at-risk ends at target cohort index date plus this value}

\item{sequenceData}{(binary) Whether to include a table at the bottom  of the plot showing the number of people at risk over time}

\item{outputFolder}{(binary) Whether to include a confidence interval}

\item{StartDays}{(string) The label for the y-axis}

\item{EndDays}{(string) The label for the y-axis}

\item{pathLevel}{(string) The label for the y-axis}
}
\value{
TRUE if it ran
}
\description{
Plot the outcome incidence over time
}
\details{
This creates a survival plot that can be used to pick a suitable time-at-risk period
}
