% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotTxPathway}
\alias{PlotTxPathway}
\title{Plot the outcome incidence over time}
\usage{
PlotTxPathway(
  cohortDatabaseSchema,
  cohortTable,
  cohortId,
  conceptSets,
  drugExposureData,
  sequenceData,
  outputFolder,
  StartDays = 0,
  EndDays = 365,
  pathLevel = 2
)
}
\arguments{
\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{cohortId}{The number of cohort id}

\item{conceptSets}{Concept set json files created from ATLAS}

\item{drugExposureData}{The covariate data of drug exposure table obtaining by FeatureExtraction package}

\item{sequenceData}{The data reformatted by drug exposure date for visualization}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{StartDays}{The first date from index date (cohort start date) for investigating drug sequence}

\item{EndDays}{The last date from index date (cohort start date) for investigating drug sequence}

\item{pathLevel}{Level of pathway depth}
}
\value{
TRUE if it ran
}
\description{
Plot the outcome incidence over time
}
\details{
This creates a survival plot that can be used to pick a suitable time-at-risk period
}
